+++ CHANGELOG.md
@@ -4,7 +4,9 @@ All notable changes to `Dezero Framework` will be documented in this file
 
 ## 0.2.7 - 2024-01-31
 
-- ArrayCollection - Created a new base class to manage arrays as an object collection
+- DataObject - Created a new implementation of concept "Data Object"
+- ArrayDataObject - New data object for arrays
+- StringDataObject - New data object for strings
 - GridView - Detect if Gridivew is loaded inside a SlidePanel
 - GridvView - Add "raw" as column default format
 - Controller - New method requireSuperadmin() for \dezero\web\Controller